Update wallet sagas to use new actions
